Move mock data to parent directory
