+++ blocks/top-table-list-block/features/top-table-list/_children/large.test.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import mockData from "./mock-data";
+import mockData from "../../../mock-data";
 
 import Large from "./large";
